worker/requirements.txt
boto3==1.35.27
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
pdf2image==1.17.0
pillow==10.4.0
python-dotenv==1.0.1
pymongo==4.8.0

